move wasm to own file; make seconds util function;
