+++ creds/src/groth16rand.rs
@@ -198,6 +198,11 @@ impl<E: Pairing> ClientState<E> {
     ) -> ShowRange<E> {
         let mut range_proof = RangeProof::default();
 
+        // prove that input is in [0, 2^n)
+        assert!(n < 64);
+        let bound = <E as Pairing>::ScalarField::from(1u64 << n);
+        assert!(ped_open.m < bound);
+
         #[cfg(feature = "wasm")]
         {
             range_proof = RangeProof::prove_n_bits(ped_open, n, &range_pk.powers);
@@ -209,11 +214,6 @@ impl<E: Pairing> ClientState<E> {
                 .num_threads(1)
                 .build()
                 .expect("Failed to create thread pool");
-
-            // prove that input is in [0, 2^n)
-            assert!(n < 64);
-            let bound = <E as Pairing>::ScalarField::from(1u64 << n);
-            assert!(ped_open.m < bound);
 
             // Use the custom thread pool for parallel operations
             pool.install(|| {
