+++ creds/src/utils.rs
@@ -8,7 +8,7 @@ use ark_std::rand::thread_rng;
 use merlin::Transcript;
 use num_bigint::BigUint;
 use sha2::{Digest, Sha512};
-use std::fs::OpenOptions;
+use std::{fs::OpenOptions, time::{SystemTime, UNIX_EPOCH}};
 use ark_std::{io::BufWriter, io::BufReader, fs::File};
 use ark_serialize::Write;
 
@@ -191,6 +191,20 @@ where
 pub fn string_to_byte_vec(s : Option<String>) -> Option<Vec<u8>> {
     // Retruns None if s is None
     Some(s?.as_bytes().to_vec())
+}
+
+pub fn utc_now_seconds() -> u64 {
+    let time_sec;
+
+    #[cfg(feature = "wasm")]
+    {
+        time_sec = crate::wasm_lib::js_timestamp();
+    }
+    #[cfg(not(feature = "wasm"))]
+    {
+        time_sec = SystemTime::now().duration_since(UNIX_EPOCH).unwrap().as_secs();
+    }
+    time_sec
 }
 
 #[cfg(test)]
